use usage filter from db

+++ app/src/main/java/com/example/ourpact3/db/UsageFiltersDao.java
@@ -22,7 +22,7 @@ public interface UsageFiltersDao {
     List<UsageFiltersEntity> getAllUsageFilters();
 
     @Query("SELECT * FROM usage_filters WHERE id = :id")
-    UsageFiltersEntity getUsageFilterById(int id);
+    UsageFiltersEntity getUsageFilterById(long id);
 
     @Update
     void updateUsageFilter(UsageFiltersEntity usageFilter);
